pyproject.toml
[tool.poetry]
name = "url-shortener"
version = "0.1.0"
description = "Project created as a recruitment task."
authors = ["Filip <mularskif@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.1"
djangorestframework = "^3.14.0"
gunicorn = "^20.1.0"
django-environ = "^0.10.0"
drf-yasg = "^1.21.5"
psycopg2 = "^2.9.6"
pytest-django = "^4.5.2"
pytest-cov = "^4.0.0"
faker = "^18.7.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
